requirements.txt
numpy
pandas
matplotlib
pillow
lxml
scikit-learn
scikit-image
jupyter
ipykernel
torch
torchvision
torchaudio
